requirements/production.txt
certifi==2022.9.24
charset-normalizer==2.1.1
docker==6.0.0
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
packaging==21.3
pyparsing==3.0.9
requests==2.28.1
smmap==5.0.0
urllib3==1.26.12
websocket-client==1.4.1
